LPS-46882 Added liferay container adapter for arquillian
This library instructs arquillian how to communicate with liferay
through arquillian plugin deployer and allows the test writer to request
for OSGi services to be injected into the test instance.

This extension also provides an adaptor that allows to modularly add
ExecutionTestListeners using arquillian extension mechanism.

So far this extension is for remote deployment of bundles only. It does
not manage containers (yet...)